example for applications
